13:05

+++ Eksamensprojekt/Eksamensprojekt/Globale Variabler.vb
@@ -1,6 +1,6 @@
 ﻿Module Globale_Variabler
     'Vi opretter nogle globale variabler som vi kan bruge til at overføre værdier fra en form til en anden.
-    Public vinkelA, vinkelB, vinkelC, sideA, sideB, sideC As Decimal 'Variabler for alle vinklerne og alle siderne
-    Public A, B, C, a1, b1, c1 As Decimal 'Variabler til vores udregninger, store bogstaver til Radianer for vinkler og små til siderne
+    Public vinkelA, vinkelB, vinkelC, sideA, sideB, sideC As Double 'Variabler for alle vinklerne og alle siderne
+    Public A, B, C, a1, b1, c1 As Double  'Variabler til vores udregninger, store bogstaver til Radianer for vinkler og små til siderne
 
 End Module
